caution edge uses near+far values
Using just near would have only worked if the lines "face" each other. This cannot be expected. It may be useful to come back to this and see if it can be justified to create a cheaper version for lines that do face each other, but I don't see it now with the expectation of very short line lengths.